pyproject.toml
[project]
name = "transcriber"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "fireworks-ai>=0.15.15",
    "google-cloud-storage>=3.10.1",
    "loguru>=0.7.3",
    "mongoengine>=0.29.3",
    "opentelemetry-exporter-otlp>=1.41.1",
    "opentelemetry-instrumentation-pymongo>=0.62b1",
    "opentelemetry-instrumentation-requests>=0.62b1",
    "pyaudio>=0.2.14",
    "pynput>=1.8.1",
    "pyperclip>=1.11.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "ruff>=0.15.9",
    "websocket-client>=1.9.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.9",
]

[tool.ruff.lint]
select = ["I", "F401"]
